Schedule Master
The Master Schedule System is a calendar application which aids the user
in making appointments, and sorting it depending on the date and time.
First the program asks the user how many appointments they would like to
make. Then, the program loops the scanning of the users input for the
name of the appointment, date and time that many times before it ends
the loop.
Next, the program prints the user’s responses in an ordered fashion. The
earliest date and time is printed out first. This is going to be done by
using bubble sort.
After it is sorted, the sorted array is written in the text file.

Methods that will be used:
Java input/output will be used by reading/writing text files.
Array/ArrayLists will be used.
Use of Bubble Sort.
Calling methods from different classes.